backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy==2.0.29
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
redis==5.0.3
aioredis==2.0.1
python-multipart==0.0.9